+++ database/migrations/2026_04_11_154707_create_dinamis_variables_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('nama_variable');
             $table->string('jenis');
-            $table->integer('variable_value');
+            $table->float('variable_value');
             $table->text('keterangan');
             $table->boolean('isactive')->nullable()->default(true);
             $table->timestamps();
